+++ tests/services/base/test_base_mail_service.py
@@ -70,42 +70,42 @@ class BaseEmailServiceTest(TestCase):
 
         self.assertIs(msg.connection, connection)
 
-    def test_get_recipient_email_list_regular(self):
+    def test_get_recipient_emails_regular(self):
         service = BaseEmailService(recipient_email_list=["dummy@example.com"])
 
-        self.assertEqual(service.get_recipient_email_list(), ["dummy@example.com"])
+        self.assertEqual(service.get_recipient_emails(), ["dummy@example.com"])
 
-    def test_get_recipient_email_list_not_set(self):
-        self.assertEqual(BaseEmailService().get_recipient_email_list(), [])
+    def test_get_recipient_emails_not_set(self):
+        self.assertEqual(BaseEmailService().get_recipient_emails(), [])
 
-    def test_build_mail_object_uses_get_recipient_email_list(self):
+    def test_build_mail_object_uses_get_recipient_emails(self):
         class MyEmailService(BaseEmailService):
             subject = "My subject"
             template_name = "testapp/test_email.html"
 
-            def get_recipient_email_list(self):
+            def get_recipient_emails(self):
                 return ["albertus.magnus@example.com"]
 
         msg = MyEmailService()._build_mail_object()
 
         self.assertEqual(msg.to, ["albertus.magnus@example.com"])
 
-    def test_is_valid_uses_get_recipient_email_list(self):
+    def test_is_valid_uses_get_recipient_emails(self):
         class MyEmailService(BaseEmailService):
             subject = "My subject"
             template_name = "testapp/test_email.html"
 
-            def get_recipient_email_list(self):
+            def get_recipient_emails(self):
                 return ["albertus.magnus@example.com"]
 
         self.assertIs(MyEmailService().is_valid(), True)
 
-    def test_is_valid_get_recipient_email_list_returns_ill_formatted_email(self):
+    def test_is_valid_get_recipient_emails_returns_ill_formatted_email(self):
         class MyEmailService(BaseEmailService):
             subject = "My subject"
             template_name = "testapp/test_email.html"
 
-            def get_recipient_email_list(self):
+            def get_recipient_emails(self):
                 return ["no-email"]
 
         with self.assertRaisesMessage(